update to welcome message, add message when changing primary
