+++ server/message_hooks.go
@@ -608,11 +608,6 @@ func (p *Plugin) SendChat(srcUser string, usersIDs []string, post *model.Post, c
 	return newMessage.ID, nil
 }
 
-func (p *Plugin) handlePromptForConnection(userID, channelID string) {
-	message := "Some users in this conversation rely on Microsoft Teams to receive your messages, but your account isn't connected. "
-	p.SendConnectMessage(channelID, userID, message)
-}
-
 func (p *Plugin) Send(teamID, channelID string, user *model.User, post *model.Post) (string, error) {
 	parentID := ""
 	if post.RootId != "" {
